third_party/WebKit/LayoutTests/installedapp/getinstalledrelatedapps-empty.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/installedapp-test-helper.js"></script>
<script>

installedapp_test((t, mock) => {
  // With no manifest, expect the Mojo call to FilterInstalledApps to be empty,
  // and return an empty list.
  mock.pushExpectedCall([], []);
  return navigator.getInstalledRelatedApps().then(result => {
    assert_array_relatedapplication_equals(result, []);
  });
}, 'getInstalledRelatedApps with no apps');

</script>
